requirements.txt
agavepy==0.7.3
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
backports.ssl-match-hostname==3.5.0.1
behave==1.2.6
Cerberus==1.2
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cloudpickle==0.6.1
configparser==3.5.0
cryptography==2.4.2
decorator==4.3.0
future==0.17.1
geneflow==1.2.0
gitdb2==2.0.5
GitPython==2.1.11
idna==2.8
iniherit==0.3.9
Jinja2==2.7.3
MarkupSafe==1.1.0
more-itertools==4.3.0
networkx==2.2
parse==1.9.0
parse-type==0.4.2
pluggy==0.8.0
py==1.7.0
pycparser==2.19
PyMySQL==0.9.2
pytest==4.0.1
python-dateutil==2.7.5
python-slugify==1.2.6
PyYAML==3.13
regex==2018.11.22
requests==2.21.0
requests-toolbelt==0.8.0
six==1.12.0
smmap2==2.0.5
SQLAlchemy==1.2.14
text-unidecode==1.2
Unidecode==1.0.23
urllib3==1.24.1
websocket-client==0.54.0
yoyo-migrations==6.0.0
